backend/graph/pyproject.toml
[project]
name = "giga_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "fastapi>=0.116.1",
    "genson>=1.3.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "joblib>=1.5.1",
    "kaleido==0.2.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-gigachat==0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.3",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-sdk>=0.2.0",
    "markdownify>=1.1.0",
    "matplotlib>=3.10.5",
    "pandas>=2.0.0",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydub>=0.25.1",
    "pytz>=2023.3",
    "psutil>=6.1.0",
    "python-dotenv>=1.1.1",
    "redis>=6.3.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "tinkoff-investments>=0.2.0b116",
    "typing-extensions>=4.14.1",
]

[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["giga_agent"]
